server/auth: scoring for offline users must load history from DB
This resolves an elusive and long-standing issue where users would 
seemingly not have their score offset by match successes in their 
history. The cause is that offline users do not have entries in the 
AuthManager's matchOutcomes and preimageOutcomes maps, but 
registerMatchOutcome and registerPreimageOutcome, which are often called 
when a user is offline, handle the lack of map entries by creating fresh 
histories that just included the outcome being registered at the time.

This PR resolves the issue by loading the user's match and preimage 
history from DB when an outcome is registered while they are offline and 
these outcome tracking map entries are non-existent. This involves some 
minor refactoring to the loadUserScore method to separate the DB loading 
component (loadUserOutcomes) from score computation (integrateOutcomes) 
and insertion of outcome map entries. I considered just deferring 
account scoring until the next login, but it is important for unbooking 
orders of suspended users.

This also modifies handleConnect so that an account will automatically 
be reinstated if computed user score indicates their account should not 
be closed but it is. This is not just for bug recovery, but to allow 
operator changing the ban score threshold without having to manually 
apply the change to suspended accounts.

This also modifies the add methods of both latest outcome tracking 
structures so that it is not possible to add duplicate matches or 
orders. This is done to ensure concurrent login/connect and match or 
swap outcome registration cannot create duplicate entries. With small 
outcome slices (~100 elements) this dumb search is inexpensive.