dex,server: lot size and rate step are market params
msgjson.Market now includes LotSize and RateStep.

msgjson.Asset still has LotSize and RateStep, but they are tagged with
omitempty, and will be zero/omitted when market configs require
multiple different lot sizes or rate steps for an asset.

dex.Asset no longer has LotSize or RateStep fields.

The dcrdex server hanress uses a configuration with two markets
with DCR as the base asset and *different* lot sizes.

The dcrdex server app now errors if markets.json has lot size or rate
step set for any asset, and if any market lacks a non-zero lot size or
rate step.

server/dex.AssetConf renamed LotSize and RateStep to LotSizeOLD and
RateStepOLD, and added omitempty to both.

server/dex.NewDEX generates a 'config' response in different ways
depending on if a given asset has consistent lot sizes and rate steps
across all markets.  If consistent, it will set the LotSize/RateStep in
the msgjson.Asset.  If not consistent, the values will be zero/omitted
and a warning will be logged that old clients will not work with the
asset.

While here in NewDEX, abort in a defer with a ready bool.