assets: use script versions meticulously
gettxout version is the transaction version, not script, and there is no
GetTxOutResult.ScriptPubKey.Version in dcrd 1.6. So, for contract audit
with dcrd 1.6 we have no choice but to pull blocks to get the raw transaction
if the transaction is confirmed. If it is still unconfirmed we can use
getrawtransaction.

harness beta node now does not enable txindex.