msgjson: bond routes and types
This adds new routes to support bonds: PostBondRoute ('postbond'),
PreValidateBondRoute('prevalidatebond'), BondExpiredRoute
('bondexpired'), and TierChangeRoute('tierchange').

This also adds the new Bond type, and updates the ConnectResult type
with the ActiveBonds slice. The ConnectResult now also has a Tier
(*int64) and LegacyFeePaid (*bool) field. The Suspended field is
deprecated.

This also adds the new BondAsset type, which specifies an accepted bond
in terms of the asset ID, bond version, required confirmations, and
increment in units of the asset's atomic amount. The ConfigResult is
updated with the BondAssets field, and a temporary BondExpiry field
to dictate expected minimum bond lock time until it is established
by consensus at BondAPIVersion. The reg fee fields are deprecated.

The following payloads are added:
- BondExpiredNotification, server-originating, when a bond expires
- TierChangedNotification, server-originating, for any tier decrease,
  including as a result of conduct violation
- PreValidateBond, client-originating, optional, with response type
- PostBond, client-originating, for adding bond or creating a new
  account with a bond, with response type