server/auth: postbond, tiers, bond tracking
This updates AuthManager with handlers for the 'postbond' and
'prevalidatebond' comms routes. New accounts may be created via
postbond, or existing accounts may have new bonds activated.
The old registration route handlers are now deprecated.

This updates the AuthManager and Market to use a new account tier
system that is a balance of active bonds and conduct score ("strikes")
instead of a hard account suspension after accumulating too many
strikes. Active bonds are monitored, and when they reach their expiry,
they are removed and the user's account tier is adjusted.
'tierchanged' notifications are sent on such events.

The (*AuthManager).Suspended method is replaced with AcctStatus,
which returns if the account is connected and current tier.

This also removes the server/account/pki package, moving its few types
and constants into the server/account package where they were used.

server/market: The methods for processing new orders now use AcctStatus
to check if an order may be accepted for a given account (tier>0).