server/{admin,dex,cmd/dcrdex}: bond assets and integ
This updates server/dex.DEX, the main DEX subsystem manager, to
recognize bond assets specified in markets.json, and prepare
the Backends as a Bonder interface that powers the AuthManager's
bond validation needs in the bond comms handlers.

This introduces a new server/dex.BondAPIVersion constant (1) that
will succeed PreAPIVersion as the first non-zero API version. We
will advance APIVersion to BondAPIVersion when the "V0PURGE" is
actuated. That will be the first stable API version, beyond which
any breaking changes will require a version bump. As suggested by
reviewers, we should probably introduce a slice of versions to
smooth out future version advances.

This creates the dex.BondExpiry(dex.Network) helper for retrieving
the bond expiry duration (in seconds) for a given network. This or
the constants BondExpiry{Mainnet,Testnet,Simnet} should become the
authoritative source for bond expiration durations for both client
and server. At this stage however, only server should reference
these compiled-in values, while the client should check the server's
'config' response and bondAssets map.

This also removes the server's "anarchy" setting, which is a relic
from early development.