client/core: fix bookie.OrderBook race
bookie embedded the orderbook.OrderBook type, which when assigning to it
as a field like b.OrderBook = orderbook.OrderBook{} would overwrite each
original field rather than treating OrderBook it as a field itself that
could safely be replaced.

This changes bookie so that it embeds a *orderbook.OrderBook, which
should force assignment to b.OrderBook to copy a new object rather than
copy each field into the existing object.