multi: Log backup swap refund txs
Add signed refund txs to the Receipt when swapping and log so that the
user can salvage funds in the case that they lose access to the client
but still have the logs. These are for BACKUP ONLY as the client will
do the refund, and if the user uses these too soon, they will cause issues.