drop authenticated upgrades
Instead of creating a full-fledged system for authenticated upgrades,
perform an upgrade of the database structure via a standard DB upgrade,
but with a sparsely populated skeleton for the stored PrimaryCredentials.
Core recognizes this condition during Login and updates the DB.
We are relying on an assumption that even if future DB upgrades are
expecting a particular structure, they won't necessarily require that
the PrimaryCredentials fields themselves are populated. This is
probably safe, since the PrimaryCredentials themselves are pretty
worthless without a user password in hand.