ui: Bin order tables by rate
This updates the order tables on the market UI. Instead of displaying each
individual order's quantity and rate on a separate row, the total quantity
available at each rate will be displayed. Also, as discussed in matrix, epoch
orders will be binned along with all other orders, and the check mark will
no longer be displayed.