client/core: set LotSize and RateStep with older servers
For newer clients that expect lotsize and rate step to be market parameters,
to still have older servers compatibility, this detects an older server by a zero
lotsize or rate step in the msgjson.Market object.  If an older server is detected,
the lot size and rate step are taken the relevant msgjson.Asset objects.

This code can be removed when support for older servers is no longer needed.