server: Allow cancel orders while market is suspended
This adds the ability for clients to submit cancel orders while a market is suspended.

The database entries after a cancel order during a market suspension will be
identical to one that gets matched while the market is running, except the
preimage will not be recorded (or requested). Also there will be no entry in the
epochs table for the epoch in which this order was "matched".