context-enable asset.Backend.FeeRate
With a websocket RPC client with auto-reconnect where RPCs hang when
the RPC server is disconnected, it is helpful to allow the caller to
cancel requests. Give FeeRate a context arg. Unfortunately, a Backend
can't cancel RPCs on shutdown however.

Try a different solution to the Synced method: add an internal 2 second
timeout. This minimizes API changes and continues to allow context
cancellation *by the Backend subsystem shutdown*, which using an
external Context precludes.

Which solution remains to be seen, but I want to limit API changes for
now, while preventing indefinite blockage when the RPC server becomes
unreachable.

Also modify server/dex.feeFetcher with the context arg, and add a
LastRate method for consumers to use as a fallback if FeeRate fails.
Fix an issue where a zero rate would be cached on FeeRate error.