rpc: Add appseed endpoint
appseed is able to return the app's seed as hex if supplied the correct
application password.

While here, fix error checking of type tests to not ignore expected
errors when no error happens.

Bump the rpcserver minor version to 1 because this adds a feature that is
backwards compatible because it is just a new endpoint. We should be
bumping the rpcSemver* vars when we make such changes.