dex/testing/dcr: connect voting wallet to beta node during reorg
During reorg, a clone of the alpha wallet is connected to the beta
node after the beta node is disconnected from the alpha node. The
added wallet uses the same seed as the alpha wallet and is thus able
to vote on tickets purchased by the alpha wallet when the alpha node
was connected to the beta node.

Without a voting wallet connected to the beta node, it can only mine 1
block after it is disconnected from the alpha node. Moreover, connecting
just any voting wallet to the beta node does not suffice as the added
wallet may not have enough tickets called to vote on a previous block
because the tickets called to vote are selected from a pool of tickets
purchased by both the new and the alpha voting wallets and the alpha
voting wallet may have more tickets called to vote on the previous block
than the other voting wallet connected to the beta node.