multi: add ability to persist password on UI login
This adds a checkbox to the login page that gives the user the option to 
persist their password for the session. When selected, all the password 
fields, other than disable account, will no longer be displayed.

When the checkbox is selected, the password is encrypted with a random 
key, and the encrypted password is stored in a cache in Core. The key to 
decrypt the password is put into a cookie. Whenever a request that 
requires a password is made with an empty password field, the webserver 
uses the key from the cookie to retrieve the password from core.