server/eth: Implement InitTxSize.
Implement InitTxSize and InitTxSizeBase.

The size of a tx does not actually tell us how much a transaction paying
to a contract with extra data will cost to send. The evm will determine
how much gas will be used. The exact cost can probably be determined but
will take a lot of digging, and will change if the contract changes. For
now, use an approximation of used gas based off of tests.

Our current implementation also does not allow for batch transactions.
As such, return zero for InitTxSizeBase in order to ensure fee
calculations equate to what we expect, which is a one for one
initialization and gas fee.