ui: Fix max order UI
There were multiple issues with the max order UI. There would be a console
error saying the the maxLoaded function was null. This would happen if there
were two concurrent requests where the one that started first finished last.
Also, if a cached value was input while a another value was being fetched,
the cached result would be overridden by the value being fetched from the
client causing a mismatch between the rate in the input box and the result
on the screen.