add topic IDs to notifications
Subjects are not language-dependent, so using the Subject field of
Notification to dictate logic is not longer sound. This work adds
a "topic" to notifications, which is a language-independent unique
ID. The topic takes the place of the subject in the core translation
maps. Additionally unexports the Translation and its fields.
Client db is upgraded to store the topic, but upgraded clients
will have no topics on pre-upgrade notifications retrieved using
(*DB).NotificationsN. This is not a problem for us anywhere, and it
would be bad practice to do use historical notifications to trigger
sensitive operations anyway, so I suspect this won't be a problem for
any existing core consumers.