server/db: external key derivation, add fee_asset
Add FeeAsset uint32 to the db.Account type and accounts table.

HD key derivation and address generation is no longer the role of the
database driver.

Add the db.FeeKeyIndexer interface, which retrieves and records an
integer index for a []byte, a serialized pubkey corresponding to
a master pubkey's child and it's index.
DEXArchivist is now required to be a FeeKeyIndexer.

CreateAccount now requires asset ID and reg address provided as inputs
rather than it generating and returning a reg address.
AccountRegAddr now also returns asset ID with the address.