dex/networks/btc: recognize rpcconnect
The bitcoin-cli setting, rpcconnect=<IP>, should override any host from
rpcbind or the default.

checkConfig is already overwriting RPCBind with the selected host and
port from a combination of defaults and RPCPort, and now RPCConnect,
so also patch RPCPort if the used port is taken from RPCBind.
This consistent approach is now documented on the Config struct.