client/core: support multiple HD accounts for same DEX
Add ability to derive additional HD account keys for a DEX
to handle cases where a user's DEX account is suspended.

Also retire the legacy registration that uses random account
keys that are not recoverable from the app seed.

Both changes ensure that all accounts registered after the
introduction of the app seed feature are recoverable from
the app seed.

Co-authored-by: Jonathan Chappelow <chappjc@protonmail.com>