networks/eth: Add methods to decode tx data.
In order to check data in a mempool transaction is formatted as we
expect and in order to read its arguments, add package methods that are
able to parse and validate Initiate and Redeem transaction data.

Some of this code is taken directly from the go-ethereum repository
where it is unexported there. abi.go and abi_test.go should be updated
if they change over there.