modules: handle auto-revocation agenda changes and update requires
Use the new major versions: dcrjson/v4 and dcrec/secp256k1/v4.

Update all other decred requires.

This introduces a breaking change to stake.DetermineTxType, which now
requires a third argument, a bool indicating if the auto-revocation
agenda is active.  We are able to say "false" always for this because
post-activation revocations ADDS constraints, letting both types of
revocations be detected.