differentiate between rate conversion factors
There are actually three rates encodings of interest. One is the
atomic ratio, which is quoteQty/baseQty. Quantities for this rate
are in atomic units. Then there is the message-rate encoding, which
is roughly uint64(atomicRatio * 1e8). Then there is the
conventional ratio, which is similar to the atomic ratio, but in
conventional units. The relationship between the atomic ratio and
the conventional ratio is

conventionalRatio = atomicRatio * (baseConversionFactor / quoteConversionFactor)

where the conversion factors are the factors for the conventional
units.

Users will want to work with the conventionalRatio, but we'll
always prefer to encode as message-rate. SimpleOrder is updated to
reflect that.