app: add historical market data chart
Adds a candlestick chart to the markets view.

server/db: Candle types moved from /server/db to /dex/candles.

core: Synchronize candlesticks and serve via BookFeed. BookFeed is changed
from a concrete type to an interface, and changes are made to improve
the design of the underlying core type. bookie now handles a map of
*candles.Cache.

app: DepthChart now inherits a new parent class, Chart, as does the
new CandleChart.

webserver: New wsLoadCandles method handles the 'loadcandles'
websocket request by finding (or creating new) the BookFeed and calling
it's Candles method. This causes candlestick data to be sent on the
BookFeeds update channel, and subsequently updates for that bin width.