pg: remove old pseudo-upgrade
A pseudo-upgrade was added prior to the first release and prior to
schema versioning that would create the matches.forgiven column
if it does not exist.  It has existed in the CREATE TABLE statement
since first release so this ALTER TABLE is always a no-op that just
results in a logged NOTICE:

[DBG] DB: pq: NOTICE (42701) - duplicate_column: column "forgiven" of relation "matches" already exists, skipping

The comments around this indicate the same, that this should have been
removed when actual upgrades were implemented. Schema v1 does
indeed include this forgiven column, and this change can thus remove
the ALTER TABLE statement that tries to re-add it on every startup.