client/eth: FundOrder, ReturnCoins, FundingCoins
This PR implements locking in the ETH wallet. Since there are no
"coins" in the ETH account system as there are in UTXO based coins,
the amount of ETH that has been reserved for orders is kept track of
in the ETH wallet.
This PR also includes a refactoring of the CoinID code in the server.
The changes in the client code needed a new coin id type which represented
coins in an address that had not yet been spent.