server/eth: Implement FundingCoin.
A funding coin represents value on the blockchain. For utxo this is a
concrete value, but for account based coins it is nebulous and changing.
For ethereum, the funding coin is represented by an address or account
and a value that account must have.

Authorization of the use of those funds is very similar to other coins,
but the user only need prove they own the keys for one account or
address.