eth/client/core: Derive account from app seed
asset.Driver is now updated with a Create function, which is
used to initialize a seeded wallet. The wallet seed is derived
from the app seed using blake256(appSeed| asssetID) and is passed
to the create function. The private key for the ETH wallet is
derived from this seed using the following BIP32 derivation path:
m/44'/60'/0'/0/0.