client: bitcoin spv
Add a built-in Bitcoin SPV wallet. Uses neutrino and btcwallet.

client/asset:

Add Create and Exists methods to the Driver. Create is used to
create a "seeded" (native, built-in) wallet, and should
not be used for non-seeded wallets. Exists checks the existence of
a wallet. Exists is written so far to check even RPC wallet
existence, though it's only used for seeded wallets in production.

Adds the WalletDefinition type. WalletInfo has a new field,
AvailableWallets []*WalletDefinition.

WalletConfig type has a new field, Type string, which should match
the Type field of one of the AvailableWallets. Existing backends
must handle Type == "" for pre-existing wallets.

client/asset/{dcr,ltc,bch,eth}: minor adaptations to fit the
Driver and WalletDefinition changes.

client/asset/btc:

Quite a few changes to configuration parsing to more clearly
separate RPC config from wallet settings. New spvWallet type
manages the *btcwallet.Wallet and *neutrino.ChainService. The
Wallet interface was already defined, so the work was just
implementing the methods using direct calls to wallet methods.

Everything is run through an interface for testing. Existing tests
required relatively few modifications to work with spv as well.
There are just a couple of new unit tests specific to the more
nuanced spv methods. The regnet_test.go/livetest.go test is adapted
to run an all-SPV trio for one test.

client/core: Work with wallet definition and give special handling
to seeded wallets.

front-end:

NewWalletForm and WalletConfigForm updated to work with
wallet definitions. Wallet type is selected via a tab at the top of
the NewWalletForm. Seeded wallets take no user-supplied password
(enforced in core). Special handling of headers and options so that
creating a seeded wallet can be very close to a 1-button, 1-click
experience.

A new form is added after the new wallet form conditionally if the
wallet balance is lower than required for the fee. This will always
be the case for the first native spv wallet. The new form shows a
deposit address and the current balance, and directs the user to
fund the address for registration fees. The balance is checked in
a loop and progresses back to the fee confirmation form when the
balance is sufficient. When working on simnet, just fund it from
the alpha node in harness-ctl.

client/db: Store the new Type field as part of the wallet. No actual
upgrade is needed.  We may consider bumping the scheme
version though.

Log btcwallet and neutrino output to a rotating log file.