eth/dex/client: Add initiateBatch function
This PR adds a batch initiation function to the ETH swap contract. This function
saves a bit of gas when initiating multiple swaps simultaneously, and also allows
ETH to conform to the current wallet interface which requires swaps to be
submitted together.  This saves ~20k gas for each additional swap that is batched.