server/asset/btc: require zpub, but offer xpub conversion
For a server operator to accept BTC for the registration fee, it is 
currently necessary to provide an extended public key with the "xpub" 
prefix. This is because the btcutil/hdkeychain package only recognizes 
such xpub strings as being from a given network, otherwise the 
net/version bytes are unknown. However, the derived fee addresses are 
P2WPKH.

The potential trouble is that wallets that generate native segwit 
addresses tend to return "zpub"-prefixed extended public key encodings 
in following BIP-84 (vpub for testnet and regnet). I have verified that 
Electrum does this if the wallet was created as native segwit, and 
Exodus actually exports both an xpub and zpub that do not seem to 
correspond to the same extended key (two separate derivation paths) 
making it potentially ill-advised to use the xpub for segwit addresses!

Prior to this PR, an operator with a wallet that provides a zpub would 
have to convert to an xpub. This is not hard (see here), but it is 
confusing and unnecessary since the dex software can do it. Further, if 
they got an xpub from a non-segwit wallet or non-segwit path, it's 
possibly they will have to go to some trouble to claim funds sent to 
segwit addresses derived from this extended key.

To address this, this PR now requires the key to be with a "zpub" 
prefix. Also, if an "xpub" is provided, it will convert it to "zpub" and 
return an error with this string so the operator is aware that segwit 
addresses will be generated if they decide to use that zpub.