client/db: active and archived matches buckets
This redefines the v5 client db upgrade to introduce a new active
matches bucket instead of creating a Retired flag on each match.

The db and upgrade code is modeled after the active/archived
orders buckets changes.

client/db.MatchIsActive is updated to detect cancel matches as
inactive. This is detected as MatchComplete and
len(proof.Auth.InitSig) == 0 since cancel matches begin as complete
and have no init/redeem data set in the Auth field.
The Address field is only empty for the maker/trade match with a
cancel, but for the taker/cancel's match with the trade it has to
this point contained the canceled maker order's address.

This also updates client/core to clear the Address field for both
sides of a cancel match. Removing this useless data from a MetaMatch
will save a bit of storage space, but previous cancel matches are not
updated by the v6 upgrade.

Shut down the db tests properly.

upgrade logger, log matches upgrade outcome