client/asset/btc: syncStatus considers wallet/addrmgr sync
There is actually a two-stage sync with btcwallet-neutrino sync,
When the chain service reaches its target height, we have to start
watching the address manager's sync height.

We could just track the address manager's sync height and ignore
the chain service height, but then it could appear to be stuck at zero
percent for a while.