trade_simnet_test fixes
Allow spv wallet to sync before funding it.  Wait a bit after funding
it for it to filter the new blocks and update balance.

Add a sleepFactor that multiplies various test sleeps when the race
detector is enabled.  Unfortunately this is a very high level harness-
based tests where actual synchronization without short arbitrary
delays is very difficult.  Just modulate the delays.