server/comms: add new routes to limiter
Add the newish price feed, spots, and candles to the routeLimiter.
The "info" limiter includes the spots and candles routes.
A new shared limiter for subscription routes is created for the
order_book and the new price_feed routes.
Add the connect route with its own rates. Burst is still 100. Clients
should self limit during account discover, as we do.