client/asset: remove time from AuditContract args
The Time argument for AuditContract is not used now that the method does
not search for the contract on the network, that now being the job of
SwapConfirmations.

This replaces the Time arg with a rebroadcast bool, so that the consumer
may decide if a rebroadcast should be attempted.

This also logs the audited counterparty contract script since that is
essential if manually redeeming the counterparty contract is required,
such as in the event of DB corruption.