eth/client: Swap
This implements the Swap function in the eth client wallet.
In addition, the eth client wallet no longer does gas estimations
using the eth client and uses hardcoded values instead. Gas
estimations do not vary unless the eth client is updated, and these
changed will be caught by the harness tests, so it is safe to just use
the hardcoded values.