dcr: stdscript migration
This migrates to the new txscript/v4/stdscript package with the newly tagged
txscript v4.0.0 module.

This also updates to tagged versions of the following dcrd modules:

- chaincfg/v3 v3.1.0
- edwards/v2 v2.0.2
- secp256k1/v4 v4.0.1
- dcrutil/v4 v4.0.0
- wire v1.5.0
- gcs/v3.0.0
- hdkeychain/v3.0.1

The new stdscript package lends itself to quite a bit of simplification and removal
of duplicate code, which as it turns out was a bit more fragile than necessary
w.r.t. script versions, new stake opcodes, etc.