dex/keygen: use ChildBIP32Std  for extended key derivation
This updates the GenDeepChild helper, which is used in both client/core
to setup DEX account keys from the app seed, and in client/asset/eth to
derive the account private key from the wallet seed.

This is especially in the eth key derivation, where the intent is to
follow BIP44 with the path m/44'/60'/0'/0/0, which requires generating
child keys as per BIP32 strictly.

For generating DEX account keys from the app seed, it is also good from
a standardness point of view to use strict BIP32 derivation.  The legacy
derivation is only required for Decred wallets for compatibility.