client/webserver/site: final dep updates with slash-div bootstrap fix
This is intended to be a final bump of node devDependencies for the v0.4
release.

Other than the bootstrap package, none of the updates affect the
contents of the generated site bundle since they are just the tooling
for webpack.

Of note is the update of bootstrap from 4.6.0 to 4.6.1, since this
includes a fix for the slew of slash-div warnings from sass dart.

Minor bumps:
- @babel/{core,eslint-parser,preset-env}: 7.15.0 -> 7.16.5
- css-loader: 6.2.0 -> 6.5.1
- css-minimizer-webpack-plugin: 3.0.2 -> 3.2.0
- eslint-plugin-import 2.24.2 -> 2.25.3
- eslint-plugin-promise: 5.1.0 -> 5.2.0
- eslint-webpack-plugin: 3.0.1 -> 3.1.1
- mini-css-extract-plugin: 2.2.0 -> 2.4.5
- sass: 1.38.1- -> 1.45.0
- sass-loader: 12.1.0 -> 12.4.0
- stylelint-webpack-plugin: 3.0.1 -> 3.1.0
- webpack: 5.51.1 -> 5.65.0
- webpack-bundle-analyzer: 4.4.2 -> 4.5.0
- webpack-cli: 4.8.0 -> 4.9.1

Patch:
- babel-loader: 8.2.2 ->8.2.3
- bootstrap: 4.6.0 -> 4.6.1

Also, clean-webpack-plugin went from "4.0.0-alpha.0" to just 4.0.0,
but these reference the same actual revision, just a new tag.